Revert "Rename files in common/LoginButton"
This reverts commit 1afbf514603aa675d2a244fcf49900a296845552.
